+++ src/CartesianBoxes.jl
@@ -25,7 +25,8 @@ export
 using Base: tail
 using Base.IteratorsMD: inc
 
-import Base: eachindex, isempty, first, last, ndims, eltype, length, size
+import Base: eachindex, isempty, first, last, ndims, eltype, length, size,
+    view, getindex, setindex!, fill!
 import Base: simd_outer_range, simd_inner_length, simd_index
 
 # Deal with compatibility issues.
@@ -113,6 +114,47 @@ axes(B::CartesianBox, d) = (first(B).I[d]:last(B).I[d])
 _dimensionlength(first::Int, last::Int) = max(0, last - first + 1)
 
 
+view(A::AbstractArray{<:Any,N}, B::CartesianBox{N}) where {N} =
+    view(A, map((i,j) -> i:j, first(B).I, last(B).I)...)
+
+function getindex(A::AbstractArray{T,N}, B::CartesianBox{N}) where {T,N}
+    empty = isempty(B)
+    empty || isnonemptypartof(B, A) ||
+        throw(BoundsError("sub-region is not part of array"))
+    C = Array{T}(undef, size(B))
+    if ! empty
+        if any(i -> i != 1, first(B).I)
+            off = CartesianIndex(map(i -> i - 1, first(B).I))
+            @inbounds @simd for i in B
+                C[i - off] = A[i]
+            end
+        else
+            @inbounds @simd for i in B
+                C[i] = A[i]
+            end
+        end
+    end
+    return C
+end
+
+setindex!(A::AbstractArray{<:Any,N}, x, B::CartesianBox{N}) where {N} =
+    A[map((i,j) -> i:j, first(B).I, last(B).I)...] = x
+
+fill!(A::AbstractArray{T,N}, B::CartesianBox{N}, x) where {T,N} =
+    fill!(A, B, convert(T, x))
+function fill!(A::AbstractArray{T,N},
+               B::CartesianBox{N}, x::T) where {T,N}
+    if ! isempty(B)
+        isnonemptypartof(B, A) ||
+            throw(BoundsError("sub-region is not part of array"))
+        @inbounds @simd for i in B
+            A[i] = x
+        end
+    end
+    return A
+end
+
+
 """
 
 `CartesianBoxable{N}` is a union of types (other than `CartesianBox{N}`) which
@@ -190,7 +232,7 @@ simd_index(iter::CartesianBox{0}, ::CartesianIndex, I1::Int) = first(iter)
 intersection(R, S)
 ```
 
-yields the Cartesian box which is the intersection of the two Cartesian boxes
+yields the Cartesian box which is the intersection of the Cartesian regions
 defined by `R` and `S`.  This method is similar to `intersect(R,S) = R ∩ S`
 which yields an array of Cartesian indices and is **much** slower (and less
 useful).
@@ -212,14 +254,14 @@ intersection(R::CartesianBoxable{N}, S::CartesianBoxable{N}) where {N} =
 isnonemptypartof(R, S)
 ```
 
-yields whether region defined by `R` is nonempty and a valid part of the region
-defined by `S` or of the contents of `S` if it is an array.
+yields whether the region defined by `R` is nonempty and a valid part of the
+region defined by `S` or of the contents of `S` if it is an array.  If this
+method returns `false`, you may call `isempty(R)` to check whether `R` was
+empty.
 
 See also: [`CartesianBox`](@ref), [`intersection`](@ref).
 
 """
-isnonemptypartof(R, S) = false
-
 function isnonemptypartof(R::Union{CartesianBoxable{N},CartesianBox{N}},
                           S::Union{CartesianBoxable{N},
                                    AbstractArray{<:Any,N}}) where {N}
@@ -238,12 +280,13 @@ isnonemptypartof(R::CartesianBox{N}, S::CartesianBox{N}) where {N} =
 boundingbox([pred,] A [, B])
 ```
 
-yields the bounding-box of values in array `A` for which the predicate function
-`pred` is true.  If the predicate function `pred` is omitted, the result is the
-bounding-box of non-zero values in array `A` or of the `true` values in `A` if
-its elements are of type `Bool`.  Optional argument `B` is to only consider a
-sub-region `B` of `A` (`B` can be a `CartesianBox`, a `CartesianIndices`, a
-`CartesianRange` or a tuple of integer unit ranges).
+yields the bounding-box of the elements in array `A` for which the predicate
+function `pred` is true.  If the predicate function `pred` is omitted, the
+result is the bounding-box of non-zero values in array `A` or of the `true`
+values in `A` if its elements are of type `Bool`.  Optional argument `B` is to
+only consider a sub-region `B` of `A` (`B` can be a `CartesianBox`, a
+`CartesianIndices`, a `CartesianRange` or a tuple of integer valued unit
+ranges).
 
 FIXME: The algorithm is pretty silly for now and could be made faster than
        `O(length(A))`.